requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
backports.csv==1.0.7
billiard==3.6.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.2
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-appconf==1.0.3
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.8.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-excel==0.0.10
django-grappelli==2.13.3
django-import-export==2.0.1
django-jsonfield==1.4.0
django-nine==0.2.3
django-rest-swagger==2.2.0
django-silk==3.0.4
django-templated-mail==1.1.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
docutils==0.16
elastic-apm==5.3.3
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.0.0
gprof2dot==2019.11.30
html5lib==1.0.1
httplib2==0.16.0
idna==2.8
importlib-metadata==1.4.0
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.3
kombu==4.6.7
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
mysqlclient==1.4.6
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
parso==0.5.2
pdfrw==0.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
protobuf==3.11.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.10.5
python-magic==0.4.15
python3-logstash==0.4.80
pytz==2019.3
PyYAML==5.3
reportlab==3.5.34
requests==2.22.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tablib==1.0.0
text-unidecode==1.3
texttable==1.6.2
tinycss2==1.0.2
traitlets==4.3.3
typed-ast==1.4.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8
WeasyPrint==51
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
zipp==2.0.0
